refactor: move layout strategies folder under new layout_handler folder
