+++ tests/chunking/strategies/merge_chunks/test_chunk_merger.py
@@ -4,8 +4,8 @@ import datetime
 
 import pytest
 
+from ingestion_pipeline.chunking.layout_handler.strategies.merge.chunk_merger import ChunkMerger
 from ingestion_pipeline.chunking.schemas import DocumentBoundingBox, DocumentChunk, DocumentMetadata
-from ingestion_pipeline.chunking.strategies.merge.chunk_merger import ChunkMerger
 
 
 # Helper function to create mock atomic chunks for tests
