+++ tests/chunking/strategies/table/test_cell_chunker.py
@@ -1,4 +1,4 @@
-"""Tests for the CellTableChunker class in ingestion_pipeline.chunking.strategies.table.cell_chunker."""
+"""Tests for the CellTableChunker class in ingestion_pipeline.chunking.layout_handler.strategies.table.cell_chunker."""
 
 from dataclasses import dataclass
 from typing import List
@@ -11,12 +11,12 @@ from textractor.entities.line import Line
 from textractor.entities.table import Table
 from textractor.entities.table_cell import TableCell
 
-import ingestion_pipeline.chunking.strategies.table.base as base_module
-import ingestion_pipeline.chunking.strategies.table.cell_chunker as cell_chunker_module
+import ingestion_pipeline.chunking.layout_handler.strategies.table.base as base_module
+import ingestion_pipeline.chunking.layout_handler.strategies.table.cell_chunker as cell_chunker_module
 from ingestion_pipeline.chunking.exceptions import ChunkException
 from ingestion_pipeline.chunking.layout_handler.layout_chunking_config import LayoutChunkingConfig
+from ingestion_pipeline.chunking.layout_handler.strategies.table.cell_chunker import CellTableChunker
 from ingestion_pipeline.chunking.schemas import DocumentMetadata
-from ingestion_pipeline.chunking.strategies.table.cell_chunker import CellTableChunker
 
 
 @pytest.fixture
