+++ tests/chunking/test_line_sentence_chunker.py
@@ -4,11 +4,11 @@ import pytest
 from textractor.entities.bbox import BoundingBox
 from textractor.entities.line import Line
 
-from ingestion_pipeline.chunking.schemas import DocumentMetadata
-from ingestion_pipeline.chunking.strategies.line_sentence_chunker import (
+from ingestion_pipeline.chunking.layout_handler.strategies.line_sentence_chunker import (
     LineSentenceChunker,
     LineSentenceChunkingConfig,
 )
+from ingestion_pipeline.chunking.schemas import DocumentMetadata
 
 
 @pytest.fixture
