+++ tests/chunking/test_textract_integration.py
@@ -11,9 +11,9 @@ from pydantic import ValidationError
 from textractor.entities.document import Document
 
 from ingestion_pipeline.chunking.layout_handler.layout_chunking_config import LayoutChunkingConfig
+from ingestion_pipeline.chunking.layout_handler.strategies.layout_text import LayoutTextChunkingStrategy
+from ingestion_pipeline.chunking.layout_handler.strategies.table import LayoutTableChunkingStrategy
 from ingestion_pipeline.chunking.schemas import DocumentBoundingBox, DocumentMetadata
-from ingestion_pipeline.chunking.strategies.layout_text import LayoutTextChunkingStrategy
-from ingestion_pipeline.chunking.strategies.table import LayoutTableChunkingStrategy
 from ingestion_pipeline.chunking.textract_layout_chunker import (
     DocumentChunk,
     TextractLayoutDocumentChunker,
